feat: ecs. movement (wip)
