+++ src/Apps/GameDesktop/Game.cs
@@ -1,4 +1,5 @@
 ﻿using Components;
+using Entities;
 using GameDesktop.Resources;
 using LightInject;
 using Mechanics;
