+++ src/Libs/Components/Jenny.properties
@@ -41,6 +41,7 @@ Jenny.PostProcessors = Jenny.Plugins.AddFileHeaderPostProcessor, \
                        Jenny.Plugins.ConsoleWriteLinePostProcessor
 Jenny.Server.Port = 3333
 Jenny.Client.Host = localhost
+Jenny.Plugins.ProjectPath = ..\Systems\Systems.csproj
 Jenny.Plugins.ProjectPath = ..\Entities\Entities.csproj
 Jenny.Plugins.ProjectPath = Components.csproj
 Entitas.CodeGeneration.Plugins.Contexts = Game, \
