+++ src/Libs/Components/Movable.cs
@@ -1,0 +1,7 @@
+﻿using Entitas;
+
+namespace Components;
+
+public class Movable : IComponent
+{
+}
